Updated message structure, changes to code to accommodate this

+++ PCD-PeerSim/src/com/pcd/P2PMessage.java
@@ -5,13 +5,13 @@ import peersim.core.Node;
 public class P2PMessage {
 	public Node sender;
 	public Node receiver;
-	public long transactionId;
+	public int transactionId;
 	public String type;    
     public long time;
 	
 	public Object[] body;
 	
-	public P2PMessage(Node s, Node r, long id, String t, long ti, Object[] b) {
+	public P2PMessage(Node s, Node r, int id, String t, long ti, Object[] b) {
 		sender = s;
 		receiver = r;
 		transactionId = id;
